+++ web/src/api/auth/graphql/fetch-user.generated.ts
