+++ web/src/api/auth/graphql/fetch-user.graphql
