--- web/src/pages/api/auth/fetch-user.ts
@@ -1,2 +1,0 @@
-import { fetchUserHandler } from "src/api/auth0/fetch-user";
-export default fetchUserHandler;
