--- web/src/pages/api/auth/update-email.ts
@@ -1,2 +1,0 @@
-import { updateEmailHandler } from "src/api/auth0/update-email";
-export default updateEmailHandler;
