--- web/src/pages/api/auth0.ts
@@ -1,2 +1,0 @@
-import { auth0Handler } from "src/api/auth0";
-export default auth0Handler;
